refactor: 댓글 관련 컴포넌트 구조 및 파일 이름 정리
- src/app/group/[groupId]/page.tsx에 있던 댓글 불러오기 로직 ReplyList 컴포넌트로 분할
- 댓글 관련 컴포넌트들 이름 변경 및 구조 정리
