--- src/components/atoms/reply/reply-content.tsx
@@ -1,7 +1,0 @@
-type ReplyContentProps = {
-  content: string;
-};
-
-export const ReplyContent = ({ content }: ReplyContentProps) => {
-  return <p>{content}</p>;
-};
