+++ src/components/organisms/reply/rereply-list.tsx
@@ -1,17 +1,20 @@
-import { ReplyContent } from '@/components/atoms/reply/reply-content';
-import { ReplyHeader } from '@/components/molecules/reply/reply-header';
+'use client';
+
+import { ReplyContent } from '@/components/molecules/reply/reply-content';
 import { useFetchInView } from '@/hooks/useFetchInView';
 import { useFetchItems } from '@/hooks/useFetchItems';
 import { Reply } from '@/types';
+import { useParams } from 'next/navigation';
 
 type RereplyListProps = {
   parentReplyId: number;
 };
 
 export const RereplyList = ({ parentReplyId }: RereplyListProps) => {
+  const { groupId } = useParams();
   const { data, fetchNextPage, hasNextPage, isFetchingNextPage } =
     useFetchItems<Reply>({
-      url: `/api/groups/1/replies/${parentReplyId}`,
+      url: `/api/groups/${groupId}/replies/${parentReplyId}`,
       queryParams: {
         size: 10,
       },
@@ -28,17 +31,17 @@ export const RereplyList = ({ parentReplyId }: RereplyListProps) => {
       <ul className="flex flex-col gap-5">
         {allRereplies.map((rereply) => (
           <li key={rereply.replyId}>
-            <ReplyHeader
+            <ReplyContent
+              content={rereply.content}
               writer={rereply.writer}
               createdAt={rereply.createdAt}
               replyId={rereply.replyId}
             />
-            <ReplyContent content={rereply.content} />
           </li>
         ))}
       </ul>
       {hasNextPage && !isFetchingNextPage && (
-        <div ref={ref} className="h-2 -translate-y-50 bg-red-500" />
+        <div ref={ref} className="h-2 -translate-y-50 bg-transparent" />
       )}
     </div>
   );
